nginx.conf

+++ src/fastcgi/pdf.cpp
@@ -20,8 +20,8 @@ void pdf_api::run()
             "\r\n"
         );
         do_convert();
-        size_t num_bytes_written = FCGX_PutStr( m_test.c_str(), m_test.length(), m_request.out );
-        if( num_bytes_written != m_test.length() || num_bytes_written == -1 )
+        int num_bytes_written = FCGX_PutStr( m_test.c_str(), m_test.length(), m_request.out );
+        if( num_bytes_written != (int)m_test.length() || num_bytes_written == -1 )
         {
             break;
         }
